Add check for sending friend request twice

+++ backend/crud.py
@@ -1,5 +1,6 @@
 from fastapi import HTTPException
 from sqlalchemy.orm import Session
+from sqlalchemy.exc import IntegrityError
 from datetime import date, datetime
 from helpers import *
 
@@ -203,12 +204,16 @@ def friend_request(db: Session, sender_id: int, receiver_id: int):
     sender = db.query(model_user).filter(model_user.id==sender_id).first()
     receiver = db.query(model_user).filter(model_user.id==receiver_id).first()
 
-    db.execute(models.user_user_association_table.insert().values(
-        requester_id=sender.id,
-        receiver_id=receiver.id,
-        state="PENDING"
-    ))
-    db.commit()
+    try :
+        db.execute(models.user_user_association_table.insert().values(
+            requester_id=sender.id,
+            receiver_id=receiver.id,
+            state="PENDING"
+        ))
+        db.commit()
+
+    except IntegrityError as e:
+        raise HTTPException(detail="You have already sent a friend request to this person!", status_code=500)
 
     return "OK"
 
@@ -216,35 +221,29 @@ def get_friend_request(db: Session, receiver_id: int, sender_id: int):
 
     sender = db.query(models.User).filter(models.User.id==sender_id).first()
     receiver = db.query(models.User).filter(models.User.id==receiver_id).first()
-
+    
     res = db.query(models.user_user_association_table).filter(
         models.user_user_association_table.c.requester_id==sender.id,
         models.user_user_association_table.c.receiver_id==receiver.id
     ).first()
+
     return res
 
-def accept_friend_request(db: Session, sender_id: int, receiver_id: int):
-
+def handle_friend_request(db: Session, sender_id: int, receiver_id: int, accepted: bool):
     sender = db.query(models.User).filter(models.User.id==sender_id).first()
     receiver = db.query(models.User).filter(models.User.id==receiver_id).first()
 
-    db.query(models.user_user_association_table).filter(
-        models.user_user_association_table.c.requester_id==sender.id,
-        models.user_user_association_table.c.receiver_id==receiver.id
-    ).update({"state": "ACCEPTED"})
-
-    db.commit()
-    return "OK"
-
-def reject_friend_request(db: Session, sender_id: int, receiver_id: int):
-    sender = db.query(models.User).filter(models.User.id==sender_id).first()
-    receiver = db.query(models.User).filter(models.User.id==receiver_id).first()
-
-    db.query(models.user_user_association_table).filter(
-        models.user_user_association_table.c.requester_id==sender.id,
-        models.user_user_association_table.c.receiver_id==receiver.id
-    ).delete()
-
+    if accepted:
+        db.query(models.user_user_association_table).filter(
+            models.user_user_association_table.c.requester_id==sender.id,
+            models.user_user_association_table.c.receiver_id==receiver.id
+        ).update({"state": "ACCEPTED"})
+    else: 
+        db.query(models.user_user_association_table).filter(
+            models.user_user_association_table.c.requester_id==sender.id,
+            models.user_user_association_table.c.receiver_id==receiver.id
+        ).delete()
+    
     db.commit()
     return "OK"
